+++ src/library/Pqi/Conformance/Operation/PipelineSync.hs
@@ -8,6 +8,7 @@ where
 import qualified Pqi
 import qualified Pqi as Lq
 import Pqi.Conformance.Harness
+import qualified Pqi.Conformance.Operation.PipelineSync.InterruptedWait as InterruptedWait
 import qualified Pqi.Conformance.Operation.PipelineSync.Interruption as Interruption
 import qualified Pqi.Conformance.Operation.PipelineSync.Parity as Parity
 import Pqi.Conformance.Prelude
@@ -19,6 +20,7 @@ spec adapter =
   describe "pipelineSync" do
     Parity.spec adapter
     Interruption.spec adapter
+    InterruptedWait.spec adapter
     it "collects pipelined queries per sync" \conninfo ->
       differential adapter conninfo \connection -> do
         entered <- Lq.enterPipelineMode connection